Parser/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
markdownify==0.11.6
networkx==3.1
html2text==2020.1.16
PyYAML==6.0.3
lxml==6.0.2